requirements.txt
Flask
pandas
openpyxl
python-dotenv
openai
xhtml2pdf
google-api-python-client
google-auth
